+++ src/mp4.rs
@@ -1,2 +1,2 @@
-pub mod file_reader;
+pub mod reader;
 pub mod writer;
