+++ src/obsws/input_registry.rs
@@ -39,7 +39,7 @@ fn transform_for_input(input: &ObswsInput) -> ObswsSceneItemTransform {
     }
     if let ObswsInputSettings::Mp4FileSource(ref settings) = input.settings
         && let Some(ref path) = settings.path
-        && let Ok(Some(dimensions)) = crate::mp4::file_reader::probe_mp4_video_dimensions(path)
+        && let Ok(Some(dimensions)) = crate::mp4::reader::probe_mp4_video_dimensions(path)
     {
         let width = dimensions.width - (dimensions.width % 2);
         let height = dimensions.height - (dimensions.height % 2);
