+++ src/obsws/source/mp4.rs
@@ -20,7 +20,7 @@ pub(super) fn build_record_source_plan(
         "obsws:{}:{run_id}:source:{source_key}:mp4_source",
         output_kind.as_str()
     ));
-    let availability = crate::mp4::file_reader::probe_mp4_track_availability(path)
+    let availability = crate::mp4::reader::probe_mp4_track_availability(path)
         .map_err(|e| BuildObswsRecordSourcePlanError::InvalidInput(e.display()))?;
     let source_video_track_id = availability.has_video.then(|| {
         TrackId::new(format!(
